Apply personal customizations
The core repo was critical to setting up my own nix/nix-darwin setup.
This is a huge step up from using bash scripts. However, I needed to
customize it to my needs. This commit is designed to be a single layer
on top of the base repo for easy rebasing upstream.

Someday, this may diverge significantly from the upstream, but for now,
easy rebasing is a nice-to-have.